Backend selection instead of using features.
This patch changes such that backend selection is done using a enum
instead of with cargo features, this is to make it more portable as a
binary instead of needing to recompile for every different backend.

This patch also changes structopt to clap and updates various
dependencies since it is a breaking update.